// src/app/interfaces/documento.ts
export interface Documento {
    estatus?: number;
    tipo?: string;
    archivo?: File;
    usuario?:number;
}


// curp: string;
// ine: string;
// titulo_licenciatura: string;
// acta_nacimiento: string;
// carta_ant_no_penales: string;
// carta_protesta1: string;
// carta_protesta2: string;
// carta_protesta3: string;
// carta_protesta4: string;
// carta_protesta5: string;
// curriculum: string;
// propuesta_programa: string;
// copia_certificada: string;